Gatherings page cards
Edit the preamble and add card components for each service.
