+++ web/Service.tsx
@@ -17,7 +17,7 @@ import Title from "./Title"
 import { Bulletin } from "./components/Bulletin"
 import Chat from "./components/Chat"
 import { Userbar } from "./components/Wagtail"
-import { useServicePageQuery, useServicePagesQuery } from "./generated-types"
+import { useServicePageQuery } from "./generated-types"
 
 gql`
   query ServicePage($slug: String!) {
@@ -33,20 +33,6 @@ gql`
           title
           description
           streamLink
-        }
-      }
-    }
-  }
-`
-
-gql`
-  query ServicePages {
-    services {
-      edges {
-        node {
-          slug
-          title
-          description
         }
       }
     }
@@ -195,68 +181,4 @@ const ServicePage: React.FC<ServicePageProps> = ({ user, ws }) => {
   }
 }
 
-type ServicesProps = {}
-
-const Services: React.FC<ServicesProps> = () => {
-  const { data, loading } = useServicePagesQuery({
-    variables: {},
-  })
-
-  if (!window.SETTINGS.PROD) {
-    console.log(loading, data)
-  }
-  const pages = data?.services?.edges
-  let listing
-  if (pages) {
-    listing = (
-      <>
-        <ul>
-          {pages.map((page: any, i: number) => (
-            <li key={i}>
-              <Link to={`/gathering/${page.node.slug}`}>{page.node.title}</Link>
-            </li>
-          ))}
-        </ul>
-      </>
-    )
-  } else {
-    listing = <Spinner animation="border"></Spinner>
-  }
-
-  return (
-    <Container className="mt-4">
-      <Title text="Gatherings" />
-      <h1>Sunday Gatherings</h1>
-      <p>
-        We gather every Sunday as a community to be encouraged and reminded that
-        we are not alone. Through praying, singing, sharing, learning,
-        discussing and laughing together, our Sunday gatherings connect us to
-        God and to each other, giving us a boost for the week to come!
-      </p>
-      <p>
-        Some of the values that mark our gatherings include:
-        <br />
-        <br />
-        <b>Informal:</b> Come as you are, not as you are ‘expected’ to be!
-        <br />
-        <b>Participatory:</b> Feel free to join the conversation, your voice is
-        welcome!
-        <br />
-        <b>Authentic:</b> We want it to be real, not religious!
-        <br />
-        <b>Relevant:</b> Connecting God to the real stuff of your life!
-        <br />
-        <br />
-        During COVID-19 we are meeting online only, through our website at 10:30
-        a.m. EST on Sunday mornings, every Sunday morning! Please feel free to
-        join us as a guest or <Link to="/signup">sign up</Link>. Our Ministry
-        Centre is closed, but our ministry continues. In partnership with
-        Johanne’s House, Southside Foodbank, John Howard Society, and other
-        tremendous organizations, we seek to serve the Ajax community
-        uninterrupted.
-      </p>
-      {listing}
-    </Container>
-  )
-}
-export { ServicePage, Services }
+export { ServicePage }
